+++ core/logging/src/main/java/dev/vitalcc/stukay/core/logging/AppLogger.kt
@@ -1,0 +1,13 @@
+package dev.vitalcc.stukay.core.logging
+
+interface AppLogger {
+    fun verbose(event: LogEvent)
+
+    fun debug(event: LogEvent)
+
+    fun info(event: LogEvent)
+
+    fun warn(event: LogEvent)
+
+    fun error(event: LogEvent)
+}