+++ core/logging/src/main/java/dev/vitalcc/stukay/core/logging/LogLevel.kt
@@ -1,0 +1,9 @@
+package dev.vitalcc.stukay.core.logging
+
+enum class LogLevel {
+    Verbose,
+    Debug,
+    Info,
+    Warn,
+    Error,
+}