+++ core/logging/src/main/java/dev/vitalcc/stukay/core/logging/LogSink.kt
@@ -1,0 +1,5 @@
+package dev.vitalcc.stukay.core.logging
+
+fun interface LogSink {
+    fun log(event: LogEvent)
+}